move function to prevent circular import
